SWH & RWH Update

+++ app/modules/scheduler/routes.js
@@ -3,6 +3,7 @@ var router = express.Router();
 var db = require('../../lib/database')();
 var flog = require('../welcome/loggedin');
 var formatAMPM = require('../welcome/formatAMPM');
+var dateformat = require('../welcome/dateformat');
 
 function regularSched(req, res, next){
   /*Regular Schedule of Current User, Match(session)
@@ -180,6 +181,69 @@ function specialSched(req, res, next){
       return next();
   });
 }
+function specialDay(req, res, next){
+  /*Selected Schedule Day of Current User, Match(session,params)
+  *(tblschedule)*/
+  db.query("SELECT * FROM tblspecialsched WHERE intSpecialAccNo= ? AND intSpecialID= ?",[req.session.user, req.params.schedid], function (err, results, fields) {
+      if (err) return res.send(err);
+      if(!(!results[0])){
+        var zero = "0";
+        var Hstart = results[0].tmSpecialStart.charAt(0).concat(results[0].tmSpecialStart.charAt(1));
+        var Hend = results[0].tmSpecialEnd.charAt(0).concat(results[0].tmSpecialEnd.charAt(1));
+        var Sampm, Eampm;
+        var formatDate = dateformat(results[0].datSpecialDate);
+
+        if (parseFloat(Hstart) > 11){
+            Sampm = 'PM';
+        }
+        else {
+            Sampm = 'AM';
+        }
+        if (parseFloat(Hstart) > 12){
+            Hstart = (parseFloat(Hstart) - 12).toString();
+        }
+        if (parseFloat(Hstart) == 0){
+            Hstart = '12';
+            Sampm = 'AM';
+        }
+        if (Hstart.length == 1){
+            Hstart = zero.concat(Hstart);
+        }
+
+        if (parseFloat(Hend) > 11){
+            Eampm = 'PM';
+        }
+        else {
+            Eampm = 'AM';
+        }
+        if (parseFloat(Hend) > 12){
+            Hend = (parseFloat(Hend) - 12).toString();
+        }
+        if (parseFloat(Hend) == 0){
+            Hend = '12';
+            Eampm = 'AM';
+        }
+        if (Hend.length == 1){
+            Hend = zero.concat(Hend);
+        }
+
+        results[0].Hstart = Hstart;
+        results[0].Hend = Hend;
+        results[0].Mstart = results[0].tmSpecialStart.charAt(3).concat(results[0].tmSpecialStart.charAt(4));
+        results[0].Mend = results[0].tmSpecialEnd.charAt(3).concat(results[0].tmSpecialEnd.charAt(4));
+        results[0].Sampm = Sampm;
+        results[0].Eampm = Eampm;
+
+        results[0].month = formatDate.charAt(4).concat(formatDate.charAt(5));
+        results[0].day = formatDate.charAt(6).concat(formatDate.charAt(7));
+        results[0].year = formatDate.charAt(0) + formatDate.charAt(1) + formatDate.toString().charAt(2) + formatDate.toString().charAt(3);
+        console.log(formatDate);
+        console.log(results);
+      }
+      req.specialDay = results;
+      return next();
+  });
+}
 
 function render(req,res){
   switch (req.valid) {
@@ -200,7 +264,7 @@ function editRWHRender(req,res){
     case 2:
     case 3:
       if (!req.regularDay[0]){
-        res.render('scheduler/views/invalid/nosched', {thisUserTab: req.user, regSchedTab: req.regularSched, empty: req.empty, specSchedTab: req.specialSched, emptyspecial: req.emptyspecial})
+        res.render('scheduler/views/invalid/nosched', {thisUserTab: req.user, regSchedTab: req.regularSched, empty: req.empty, specSchedTab: req.specialSched, emptyspecial: req.emptyspecial});
       }
       else{
         res.render('scheduler/views/editRWH', {thisUserTab: req.user, regSchedTab: req.regularSched, regDayTab: req.regularDay, empty: req.empty, specSchedTab: req.specialSched, emptyspecial: req.emptyspecial});
@@ -209,17 +273,65 @@ function editRWHRender(req,res){
   }
 }
 function delRWHRender(req,res){
-  db.query("DELETE FROM tblschedule WHERE intSchedID= ?",[req.params.schedid], (err, results, fields) => {
-    if (err) console.log(err);
-    res.redirect('/scheduler');
-  });
+  switch (req.valid) {
+    case 1:
+      res.render('welcome/views/invalid/adm-restrict');
+      break;
+    case 2:
+    case 3:
+      if(!req.regularDay[0]){
+        res.render('scheduler/views/invalid/nosched', {thisUserTab: req.user, regSchedTab: req.regularSched, empty: req.empty, specSchedTab: req.specialSched, emptyspecial: req.emptyspecial})
+      }
+      else{
+        db.query("DELETE FROM tblschedule WHERE intSchedID= ?",[req.params.schedid], (err, results, fields) => {
+          if (err) console.log(err);
+          res.redirect('/scheduler');
+        });
+      }
+  }
+}
+function editSWHRender(req,res){
+  switch (req.valid) {
+    case 1:
+      res.render('welcome/views/invalid/adm-restrict');
+      break;
+    case 2:
+    case 3:
+      if (!req.specialDay[0]){
+        res.render('scheduler/views/invalid/nosched', {thisUserTab: req.user, regSchedTab: req.regularSched, empty: req.empty, specSchedTab: req.specialSched, emptyspecial: req.emptyspecial})
+      }
+      else{
+        res.render('scheduler/views/editSWH', {thisUserTab: req.user, regSchedTab: req.regularSched, specDayTab: req.specialDay, empty: req.empty, specSchedTab: req.specialSched, emptyspecial: req.emptyspecial});
+      }
+      break;
+  }
+}
+function delSWHRender(req,res){
+  switch (req.valid) {
+    case 1:
+      res.render('welcome/views/invalid/adm-restrict');
+      break;
+    case 2:
+    case 3:
+      if (!req.specialDay[0]){
+        res.render('scheduler/views/invalid/nosched', {thisUserTab: req.user, regSchedTab: req.regularSched, empty: req.empty, specSchedTab: req.specialSched, emptyspecial: req.emptyspecial})
+      }
+      else{
+        db.query("DELETE FROM tblspecialsched WHERE intSpecialID= ?",[req.params.schedid], (err, results, fields) => {
+          if (err) console.log(err);
+          res.redirect('/scheduler');
+        });
+      }
+  }
 }
 
 router.get('/', flog, regularSched, specialSched, render);
 router.get('/rwh/:schedid', flog, regularSched, regularDay, specialSched, editRWHRender);
-router.get('/rwh/:schedid/remove', flog, delRWHRender);
+router.get('/rwh/:schedid/remove', flog, regularSched, regularDay, specialSched, delRWHRender);
+router.get('/swh/:schedid', flog, regularSched, specialDay, specialSched, editSWHRender);
+router.get('/swh/:schedid/remove', flog, regularSched, specialDay, specialSched, delSWHRender);
 
-router.post('/rwh', (req, res) => {
+router.post('/rwh', flog, (req, res) => {
   if (req.body.Sampm == 'AM' && req.body.Shours == '12'){
       req.body.Shours = '00';
   }
@@ -239,7 +351,7 @@ router.post('/rwh', (req, res) => {
     res.redirect('/scheduler');
   });
 });
-router.post('/rwh/:schedid', (req, res) => {
+router.post('/rwh/:schedid', flog, (req, res) => {
   if (req.body.ESampm == 'AM' && req.body.EShours == '12'){
       req.body.EShours = '00';
   }
@@ -259,7 +371,7 @@ router.post('/rwh/:schedid', (req, res) => {
     res.redirect('/scheduler');
   });
 });
-router.post('/swh', flog, regularSched, regularDay, specialSched, editRWHRender, (req, res) => {
+router.post('/swh', flog, regularSched, specialSched, (req, res) => {
   if (req.body.Sampm == 'AM' && req.body.Shours == '12'){
       req.body.Shours = '00';
   }
@@ -277,6 +389,27 @@ router.post('/swh', flog, regularSched, regularDay, specialSched, editRWHRender,
   var date = req.body.addYear.toString()+'-'+req.body.addMonth+'-'+req.body.addDay;
   db.query("INSERT INTO tblspecialsched (intSpecialAccNo, datSpecialDate, tmSpecialStart, tmSpecialEnd) VALUES (?,?,?,?)",[req.session.user, date, start, end], (err, results, fields) => {
     if (err) res.render('scheduler/views/invalid/nodate', {thisUserTab: req.user, regSchedTab: req.regularSched, empty: req.empty, specSchedTab: req.specialSched, emptyspecial: req.emptyspecial});
+    res.redirect('/scheduler');
+  });
+});
+router.post('/swh/:schedid', flog, (req, res) => {
+  if (req.body.ESampm == 'AM' && req.body.EShours == '12'){
+      req.body.EShours = '00';
+  }
+  if (req.body.EEampm == 'AM' && req.body.EEhours == '12'){
+      req.body.EEhours = '00';
+  }
+  if (req.body.ESampm == 'PM' && req.body.EShours != '12'){
+      req.body.EShours = (parseFloat(req.body.EShours) + 12).toString();
+  }
+  if (req.body.EEampm == 'PM' && req.body.EEhours != '12'){
+      req.body.EEhours = (parseFloat(req.body.EEhours) + 12).toString();
+  }
+  var start = req.body.EShours.concat(':'+req.body.ESminutes);
+  var end = req.body.EEhours.concat(':'+req.body.EEminutes);
+  var date = req.body.addYear.toString()+'-'+req.body.addMonth+'-'+req.body.addDay;
+  db.query("UPDATE tblspecialsched SET datSpecialDate= ?, tmSpecialStart= ?, tmSpecialEnd= ? WHERE intSpecialID= ?",[date, start, end, req.params.schedid], (err, results, fields) => {
+    if (err) console.log(err);
     res.redirect('/scheduler');
   });
 });
